Functions makeRow and makeGrid
The makeRow function creates a row of blocks given the parameter, x.

The makeGrid function needs to be able to take two parameters, x and y,
to run makeRow(x) and y will determine how many times the makeRow
function runs. In addition, makeGrid needs to correctly insert a break
to start a new row of blocks.